Add second-level exception continuation hook
The current enclave exception handler has a stack format that does not
allow for ocalls (rsp does not point to the correct stack frame). Add
a second-level exception handler that is called after stack state is
restored to allow the handler to make ocalls.

Additionally, add a test which makes use of this functionality.

Signed-off-by: Mike Brasher <mikbras@microsoft.com>
Co-authored-by: Jordan Hand <jorhand@microsoft.com>
Co-authored-by: Mike Brasher <mikbras@microsoft.com>